backend/requirements.txt
# Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
cloudinary==1.38.0
httpx==0.26.0
requests==2.31.0
google-auth==2.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
certifi
